limpieza, norma y comentarios para aclarar la parte de ejecución
