+++ documents/Vadim/bash_testing/ideas_tests
@@ -6,7 +6,7 @@
 #    By: vjan-nie <vjan-nie@student.42madrid.com    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/08/26 16:16:32 by vjan-nie          #+#    #+#              #
-#    Updated: 2025/09/17 13:44:59 by vjan-nie         ###   ########.fr        #
+#    Updated: 2025/09/22 12:24:22 by vjan-nie         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -148,6 +148,8 @@ stop
 echo ls -l | wc -l > outfile
 
 sh << stop | wc -l > outfile (biennnnn)
+
+ls | << stop cat | cat > outfile
 
 NUEVOS PROBLEMAS:
