+++ includes/structs.h
@@ -6,7 +6,7 @@
 /*   By: vjan-nie <vjan-nie@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/19 23:30:20 by serjimen          #+#    #+#             */
-/*   Updated: 2025/09/16 14:13:32 by vjan-nie         ###   ########.fr       */
+/*   Updated: 2025/09/22 14:57:18 by vjan-nie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
