+++ src/pipes_utils.c
@@ -6,7 +6,7 @@
 /*   By: vjan-nie <vjan-nie@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/02 15:23:06 by vjan-nie          #+#    #+#             */
-/*   Updated: 2025/09/12 11:16:01 by vjan-nie         ###   ########.fr       */
+/*   Updated: 2025/09/22 14:56:50 by vjan-nie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ void	ft_close_two(int fd1, int fd2)
 void	free_pipe_data(t_pipe *pipe_data)
 {
 	if (!pipe_data)
-		return;
+		return ;
 	if (pipe_data->in > 2)
 		close(pipe_data->in);
 	if (pipe_data->out > 2)
@@ -49,7 +49,7 @@ void	free_pipe_data(t_pipe *pipe_data)
  * containing what we need in order to execute trhough
  * pipes.
 */
-t_pipe	*init_pipe_data(t_command *command, t_env **env_list, size_t nbr_of_commands)
+t_pipe	*init_pipe_data(t_command *command, t_env **env_list, size_t commands)
 {
 	t_pipe	*pipe_data;
 
@@ -57,7 +57,7 @@ t_pipe	*init_pipe_data(t_command *command, t_env **env_list, size_t nbr_of_comma
 	if (!pipe_data)
 		return (NULL);
 	pipe_data->commands = command;
-	pipe_data->command_count = nbr_of_commands;
+	pipe_data->command_count = commands;
 	pipe_data->env_list = env_list;
 	pipe_data->in = STDIN_FILENO;
 	pipe_data->out = STDOUT_FILENO;
